isochron: use memmove in isochron_log_remove
When we overwrite a log element we use memcpy, but valgrind rightfully
complains:

==12713== Source and destination overlap in memcpy(0x4a89410, 0x4a89448, 5544)
==12713==    at 0x484F294: __GI_memcpy (vg_replace_strmem.c:1035)
==12713==    by 0x403F2F: isochron_log_remove (in /root/isochron)
==12713==    by 0x40B4DF: isochron_print_stats (in /root/isochron)
==12713==    by 0x40B897: prog_teardown (in /root/isochron)
==12713==    by 0x40C85F: isochron_send_main (in /root/isochron)
==12713==    by 0x401F2F: main (in /root/isochron)
==12713==

"man memcpy" says:

The memory areas must not overlap.  Use memmove(3) if the memory areas do overlap.

So let's use memmove.

+++ isochron/common.h
@@ -144,7 +144,7 @@ int isochron_log_recv(struct isochron_log *log, int fd);
 void isochron_log_teardown(struct isochron_log *log);
 void isochron_rcv_log_print(struct isochron_log *log);
 void isochron_send_log_print(struct isochron_log *log);
-void isochron_log_remove(struct isochron_log *log, void *p, int len);
+void isochron_log_remove(struct isochron_log *log, void *p, size_t len);
 
 enum isochron_management_id {
 	ISOCHRON_MID_LOG,
